# [USMON-547] faster protocol classification tests