--- a/pkg/network/protocols/mysql/testdata/docker-compose.yml
+++ b/pkg/network/protocols/mysql/testdata/docker-compose.yml
@@ -8,3 +8,6 @@ services:
       MYSQL_ROOT_PASSWORD: ${MYSQL_ROOT_PASS:-root}
     ports:
       - ${MYSQL_ADDR:-127.0.0.1}:${MYSQL_PORT:-3306}:3306
+    tmpfs:
+      - /var/lib/mysql
+