--- a/pkg/network/protocols/testutil/serverutils.go
+++ b/pkg/network/protocols/testutil/serverutils.go
@@ -27,6 +27,11 @@ const (
 // - serverStartRegex is a regex to be matched on the server logs to ensure it started correctly.
 func RunDockerServer(t testing.TB, serverName, dockerPath string, env []string, serverStartRegex *regexp.Regexp, timeout time.Duration) error {
 	t.Helper()
+	// Ensuring no previous instances exists.
+	c := exec.Command("docker-compose", "-f", dockerPath, "down", "--remove-orphans")
+	c.Env = append(c.Env, env...)
+	_ = c.Run()
+
 	ctx, cancel := context.WithCancel(context.Background())
 	t.Cleanup(cancel)
 
@@ -44,7 +49,8 @@ func RunDockerServer(t testing.TB, serverName, dockerPath string, env []string, 
 
 		c := exec.Command("docker-compose", "-f", dockerPath, "down", "--remove-orphans")
 		c.Env = append(c.Env, env...)
-		_ = c.Run()
+		// Not waiting for its finish.
+		_ = c.Start()
 	})
 
 	for {